enhance: Split reducer into functions for each action
