+++ packages/core/src/internal.ts
@@ -1,3 +1,3 @@
 export { inferResults, DELETED } from '@rest-hooks/normalizr';
 export { default as RIC } from './state/RIC.js';
-export { initialState } from './state/createReducer.js';
+export { initialState } from './state/reducer/createReducer.js';
