+++ packages/core/src/manager/__tests__/networkManager-legacy.ts
@@ -3,7 +3,7 @@ import { ArticleResource } from '__tests__/legacy-3';
 import { RECEIVE_TYPE } from '../../actionTypes';
 import Controller from '../../controller/Controller';
 import { FetchAction } from '../../legacyActions';
-import { initialState } from '../../state/createReducer';
+import { initialState } from '../../state/reducer/createReducer';
 import { createFetch } from '../../state/legacy-actions';
 import { Middleware } from '../../types';
 import NetworkManager from '../NetworkManager';
