+++ packages/core/src/manager/__tests__/pollingSubscription.ts
@@ -1,6 +1,6 @@
 import { PollingArticleResource } from '__tests__/legacy-3';
 
-import { initialState } from '../../state/createReducer';
+import { initialState } from '../../state/reducer/createReducer';
 import ConnectionListener from '../ConnectionListener';
 import DefaultConnectionListener from '../DefaultConnectionListener';
 import PollingSubscription from '../PollingSubscription';
