+++ website/src/components/Playground/editor-types/@rest-hooks/rest.d.ts
@@ -901,6 +901,7 @@ interface RestInstance<
   process(value: any, ...args: Parameters<F>): ResolveType<F>;
 
   /* utilities */
+  /** @see https://resthooks.io/rest/api/RestEndpoint#testKey */
   testKey(key: string): boolean;
 
   /* extenders */
